Create a new namespace litigators for the litigator hardship claim api
