+++ app/interfaces/api/v1/external_users/claim.rb
@@ -14,7 +14,7 @@ module API
           mount API::V1::ExternalUsers::Claims::FinalClaim
           mount API::V1::ExternalUsers::Claims::InterimClaim
           mount API::V1::ExternalUsers::Claims::TransferClaim
-          mount API::V1::ExternalUsers::Claims::HardshipClaim
+          mount API::V1::ExternalUsers::Claims::Litigators::HardshipClaim
         end
       end
     end
