+++ app/views/pages/api_release_notes.html.haml
@@ -21,9 +21,9 @@
       The new endpoint is interactively documented at the links below:
       %ul.list-bullet
         %li
-          %a{ href: '/api/documentation?#!/api/postApiExternalUsersClaimsHardship', target: '_blank' } litigator hardship claim creation
+          %a{ href: '/api/documentation?#!/api/postApiExternalUsersClaimsLitigatorsHardship', target: '_blank' } litigator hardship claim creation
         %li
-          %a{ href: '/api/documentation?#!/api/documentation#!/api/postApiExternalUsersClaimsHardshipValidate', target: '_blank' } litigator hardship claim validate
+          %a{ href: '/api/documentation?#!/api/documentation#!/api/postApiExternalUsersClaimsLitigatorsHardship', target: '_blank' } litigator hardship claim validate
       %br/
       %br/
       A list of valid Case Stages (to be used when creating a hardship claim) can be retrieved via:
