+++ spec/api/v1/external_users/claim_spec.rb
@@ -25,8 +25,8 @@ RSpec.describe API::V1::ExternalUsers::Claim do
     /api/external_users/claims/transfer
     /api/external_users/claims/transfer/validate
 
-    /api/external_users/claims/hardship
-    /api/external_users/claims/hardship/validate
+    /api/external_users/claims/litigators/hardship
+    /api/external_users/claims/litigators/hardship/validate
   ).freeze
 
   let(:api_routes) do
